Convert ElasticsearchSearchQueryTransformer to Byte Buddy
