+++ stagemonitor-requestmonitor/src/main/resources/META-INF/services/org.stagemonitor.core.instrument.StagemonitorByteBuddyTransformer
@@ -1,1 +1,2 @@
 org.stagemonitor.requestmonitor.profiler.ProfilingTransformer
+org.stagemonitor.requestmonitor.profiler.elasticsearch.ElasticsearchSearchQueryTransformer
