+++ stagemonitor-requestmonitor/src/main/resources/META-INF/services/org.stagemonitor.core.instrument.StagemonitorJavassistInstrumenter
@@ -1,3 +1,2 @@
 org.stagemonitor.requestmonitor.MonitorRequestsInstrumenter
-org.stagemonitor.requestmonitor.profiler.elasticsearch.ElasticsearchSearchQueryInstrumenter
 org.stagemonitor.requestmonitor.ejb.RemoteEjbMonitorInstrumenter
