+++ stagemonitor-requestmonitor/src/test/resources/stagemonitor.properties
@@ -4,4 +4,5 @@ stagemonitor.instrument.include=org.stagemonitor.requestmonitor.prof.,\
   org.stagemonitor.requestmonitor.MultipleAnnotationsAndProfilerTest,\
   org.stagemonitor.requestmonitor.MonitorRequestsInstrumenterTest$TestClass,\
   org.stagemonitor.requestmonitor.ejb.RemoteEjbMonitorInstrumenterTest$RemoteInterfaceImpl
+stagemonitor.instrument.exclude=org.stagemonitor.requestmonitor.prof.elasticsearch.ElasticsearchSearchQueryTransformerTest
 stagemonitor.businessTransaction.namingStrategy=CLASS_NAME_HASH_METHOD_NAME
